MOSIP-25416 bug fixed, key moved to oidc-server
