+++ oidc-server/config.js
@@ -1,4 +1,7 @@
 module.exports = {
-  IDP_BASE_URL: process.env.IDP_BASE_URL,
-  PORT: process.env.PORT,
+  IDP_BASE_URL: process.env.IDP_BASE_URL ?? "http://localhost:8088/v1/idp",
+  PORT: process.env.PORT ?? 8888,
+  PRIVATE_KEY: process.env.PRIVATE_KEY ?? "-----BEGIN PRIVATE KEY-----\nMIIEvQIBADANBgkqhkiG9w0BAQEFAASCBKcwggSjAgEAAoIBAQDBcZADnvgJT5ZN\nY8gt3hLf/Wwl/Rih/STAx41y1dzW2FCjvQuN+f1/iKClF7t/yo2z9bp0u4tZ6YWE\nYkFAFDu++JyQQIo0l+j8TNIP/LI8KDuT8AZtNZOS6g5seumKXAkbIpVKQ4i6Mf8W\nochWTCRVeNZzbkuxvwaricAc1ASqqBpg8sSLl68JHpLQXNxdM+5GH3Mf5wzQ6eO/\niiut9ikSvrUdxGnkxcvnlvvIGghiVcPw19qVyB2tbOnhVke9DzJP0bl5p/btdSfp\nkF4F/tWcQtwoVPLDWkrbCqg7FEFzxjXA7qsTLw+UJTrAP6ZsjNyMbXBR+duC8k4z\ngDkbjWq5AgMBAAECggEABgdeiCZbr5qZ4haShg9uQinZRYPSUTYc/58YgvQ0WkPK\nm5fINOgOJPvimdKYBt8OtIWbhojTyn0TKrGPPAqFZCnGY16HkCUN31MbluD2wxYz\n6SPpZ1zsmP8PbQUVozjEFeLpiTN6nubw/skS/9GGrl1CPb25wTPlZtI3uQ5IiPL/\nYD5j/w5/J7tejAaRbhlJj48ZDa4CR8BkaUi2QaQmLoyiO/1O+U+Nf17+t1C6zFFK\nKHQx2lNltE1xFQoHB4WuBA2GnP5LgNFJSLv0p95gQK37nP0TTcuiZVlvFcmbGI/i\nlWlxRKJUD3mZR6nz25X4SapUWswnrnm7JtUA/UGVGwKBgQDT7jQhLFcOYLxJWu52\nDV7u9CyTJaWiFp2bzdECRTIfpLiHVpM+s3AtWf1YA2fJFOq6kPUWvyZhnt/2dnYy\nQyPjlRXy6IiNBmvdysyrRgaTO38u/CEt25Y4n/P3/3Nk0dyRd6LgbxBAUlBhvdM1\n0axVVLB9/rxJKh6fyiq+RuQAmwKBgQDpqzzyh1xs/YwNP2WRqErbPhAgfbnOxr9R\n0uAC0vANWl6Rd8k05C+6o1jUbDcnCzq48MsZwQrjYzJ450IWqSUVO8MmUlcwW+Ba\nuYZsJQy8kv6sXI76QJMacUFfrSjN0luN8G3cTdBjHcb3TtmK672jRrN7LUMm2r/M\nrB5qsT7FOwKBgHt29l+tGTqDGK4Tj4Cu4XOMRLuk3w3C5ybxjKYp8b2FlQsjYLzS\nhLOnDXTM5xW6XdU6xc3upkmQ6SEIemtlue7H1/Wfio2+A/E+SGYvcX0s2DdjI2E3\n4Bg090ivRWai7EDC1dyyV2FhkvqeQAHGmQNNR1TfpolpRFYKzUP0wLepAoGAYg+B\nqUoTCI4y6xBS4JieqXlXLTt18YfInF8BsU2yffgRvbxmTPMB8LJCQgsT7iexQhGT\nOkCgACMN+F0ciAP90vZchEWD34B/G7PF7LZzrOOHSvAg9HaLBUrII424lP+VenCO\nuihRrna9m+WUN8+MquutwKCTEMg2O39z2FR/wicCgYEApmL/G7T4OF/pr2RCzkku\npi1dCbwRX39bMEIs3uirvkoPR5CENvuvsXQ0Oias3taxzLa4nG5JVXHkyOIX8UsK\n1NFrzZPRKbfNX3h5EAnl3I7cZMtoYJLnawUqaNTukOmDChPlKx1fVjUwsyNn5HSA\nnmBiaOmm/RHo36tPhgaPUtE=\n-----END PRIVATE KEY-----",
+  IDP_AUD_URL: process.env.IDP_AUD_URL ?? "http://localhost:8088/v1/idp/oauth/token",
+  CLIENT_ASSERTION_TYPE: process.env.CLIENT_ASSERTION_TYPE ?? "urn:ietf:params:oauth:client-assertion-type:jwt-bearer"
 };
