+++ oidc-server/idpService.js
@@ -1,25 +1,27 @@
 const axios = require("axios");
-const { IDP_BASE_URL } = require("./config");
+const jose = require("jose");
+const { IDP_BASE_URL, PRIVATE_KEY, IDP_AUD_URL, CLIENT_ASSERTION_TYPE } = require("./config");
 
 const baseUrl = IDP_BASE_URL.trim();
 const getTokenEndPoint = "/oauth/token";
 const getUserInfoEndPoint = "/oidc/userinfo";
 
+const alg = "RS256";
+const expirationTime = "1h";
+
 const post_GetToken = async ({
   code,
   client_id,
   redirect_uri,
-  grant_type,
-  client_assertion_type,
-  client_assertion,
+  grant_type
 }) => {
   let request = new URLSearchParams({
     code: code,
     client_id: client_id,
     redirect_uri: redirect_uri,
     grant_type: grant_type,
-    client_assertion_type: client_assertion_type,
-    client_assertion: client_assertion,
+    client_assertion_type: CLIENT_ASSERTION_TYPE,
+    client_assertion: await generateSignedJwt(client_id),
   });
   const endpoint = baseUrl + getTokenEndPoint;
   const response = await axios.post(endpoint, request, {
@@ -40,6 +42,32 @@ const get_GetUserInfo = async (access_token) => {
   });
 
   return response.data;
+};
+
+
+const generateSignedJwt = async (clientId) => {
+  // Set headers for JWT
+  var header = {
+    alg: alg,
+    typ: "JWT",
+  };
+
+  var payload = {
+    iss: clientId,
+    sub: clientId,
+    aud: IDP_AUD_URL,
+  };
+
+  // const keyObjFromJwk = await jose.importJWK(public_private_key_jwk, alg);
+  var privateKey = await jose.importPKCS8(PRIVATE_KEY, alg);
+
+  const jwt = new jose.SignJWT(payload)
+    .setProtectedHeader(header)
+    .setIssuedAt()
+    .setExpirationTime(expirationTime)
+    .sign(privateKey);
+
+  return jwt;
 };
 
 module.exports = {
