+++ oidc-ui/src/components/Registration.js
@@ -8,7 +8,6 @@ import { useTranslation } from "react-i18next";
 import RedirectButton from "../common/RedirectButton";
 
 export default function Registration({
-  cryptoService,
   clientService,
   oidcService,
   i18nKeyPrefix = "registration",
@@ -19,10 +18,6 @@ export default function Registration({
 
   const { getURIforRegistration } = {
     ...clientService,
-  };
-
-  const { generateSignedJwt } = {
-    ...cryptoService,
   };
 
   const { post_fetchUserInfo } = {
@@ -75,16 +70,12 @@ export default function Registration({
       let client_id = clientDetails.clientId;
       let redirect_uri = clientDetails.redirect_uri_registration;
       let grant_type = clientDetails.grant_type;
-      let client_assertion_type = clientDetails.client_assertion_type;
-      let client_assertion = await generateSignedJwt(client_id);
 
       var userInfo = await post_fetchUserInfo(
         authCode,
         client_id,
         redirect_uri,
-        grant_type,
-        client_assertion_type,
-        client_assertion
+        grant_type
       );
       setUserInfo(userInfo);
       setStatus(states.LOADED);
@@ -106,7 +97,7 @@ export default function Registration({
             <label>{t("full_name")}</label>
             <input
               type="text"
-              value={userInfo?.given_name}
+              value={userInfo?.given_name ?? userInfo?.name}
               className="rounded bg-gray-50 border text-gray-900 focus:ring-blue-500 focus:border-blue-500 block flex-1 min-w-0 text-sm border-gray-300 p-2.5"
             />
           </div>
@@ -129,10 +120,13 @@ export default function Registration({
             />
           </div>
           <div className="w-full flex flex-col mb-6 text-slate-500">
-            <label>{t("city")}</label>
+            <label>{t("phone_number")}</label>
             <input
               type="text"
               className="rounded bg-gray-50 border text-gray-900 focus:ring-blue-500 focus:border-blue-500 block flex-1 min-w-0 text-sm border-gray-300 p-2.5"
+              value={userInfo?.phone_number_verified ??
+                userInfo?.phone ??
+                userInfo?.phone_number}
             />
           </div>
         </div>
