+++ oidc-ui/src/pages/Registration.js
@@ -2,7 +2,6 @@ import React from "react";
 import Background from "../components/Background";
 import Registration from "../components/Registration";
 import clientService from "../services/clientService";
-import cryptoService from "../services/cryptoService";
 import langConfigService from "../services/langConfigService";
 import oidcService from "../services/oidcService";
 
@@ -12,7 +11,6 @@ export default function RegistrationPage() {
       langConfigService={langConfigService}
       component={React.createElement(Registration, {
         clientService: clientService,
-        cryptoService: cryptoService,
         oidcService: oidcService,
       })}
     />
