+++ oidc-ui/src/pages/UserProfile.js
@@ -1,7 +1,6 @@
 import React from "react";
 import Background from "../components/Background";
 import UserProfile from "../components/UserProfile";
-import cryptoService from "../services/cryptoService";
 import langConfigService from "../services/langConfigService";
 import oidcService from "../services/oidcService";
 
@@ -10,7 +9,6 @@ export default function UserProfilePage() {
     <Background
       langConfigService={langConfigService}
       component={React.createElement(UserProfile, {
-        cryptoService: cryptoService,
         oidcService: oidcService,
       })}
     />
